+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Contracts/Project/Tasks/ProjectByTaskSummaryResponse.cs
@@ -59,6 +59,7 @@ namespace Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks
         
         public TaskSummaryResponse PrincipalDesignate { get; set; }
         
+        public TaskSummaryResponse DueDiligenceChecks { get; set; }
         public int TaskCount { get; set; }
         
         public int CompletedTasks { get; set; }