+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Tasks/PDG/Central/AddPDGPayment.cshtml.cs
@@ -10,12 +10,8 @@ using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks;
 using Dfe.ManageFreeSchoolProjects.Logging;
 using System.Threading.Tasks;
 using Dfe.ManageFreeSchoolProjects.Constants;
-using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks.PDG;
 using Dfe.ManageFreeSchoolProjects.Validators;
 using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Payments;
-using System.Text.RegularExpressions;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
-using DocumentFormat.OpenXml.Drawing;
 
 namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
 {
@@ -30,8 +26,6 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
         [BindProperty(SupportsGet = true, Name = "projectId")]
         public string ProjectId { get; set; }
 
-        public int TotalGrant { get; set; }
-
         public string CurrentFreeSchoolName { get; set; }
         [BindProperty(Name = "payment-due-date", BinderType = typeof(DateInputModelBinder))]
         [Display(Name = "Date the payment is due")]
@@ -41,6 +35,7 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
 
         [BindProperty(Name = "payment-due-amount", BinderType = typeof(DecimalInputModelBinder))]
         [Display(Name = "Due amount")]
+        [ValidMoney(0, 640000)]
         [Required]
         public decimal? PaymentScheduleAmount { get; set; }
 
@@ -50,7 +45,8 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
         public DateTime? PaymentActualDate { get; set; }
 
         [BindProperty(Name = "payment-actual-amount", BinderType = typeof(DecimalInputModelBinder))]
-        [Display(Name = "Actual amount")]
+        [Display(Name = "Amount sent")]
+        [ValidMoney(0, 640000)]
         public decimal? PaymentActualAmount { get; set; }
 
         public AddPDGPaymentModel(IGetProjectByTaskService getProjectService, IAddProjectPaymentsService addProjectPaymentsService, ILogger<AddPDGPaymentModel> logger,
@@ -76,30 +72,10 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
 
             var totalGrant = project.PDGDashboard.RevisedGrant ?? project.PDGDashboard.InitialGrant;
 
-            TotalGrant = (int)totalGrant;
-
             CurrentFreeSchoolName = project.SchoolName;
 
             if (!ModelState.IsValid)
             {
-                _errorService.AddErrors(ModelState.Keys, ModelState);
-                return Page();
-            }
-
-            var validPaymentScheduleAmount = IsValidMoney(PaymentScheduleAmount, "Due amount");
-
-            if (validPaymentScheduleAmount != ValidationResult.Success)
-            {
-                ModelState.AddModelError("payment-due-amount", validPaymentScheduleAmount.ErrorMessage);
-                _errorService.AddErrors(ModelState.Keys, ModelState);
-                return Page();
-            }
-
-            var validPaymentActualAmount = IsValidMoney(PaymentActualAmount, "Actual amount");
-
-            if (validPaymentActualAmount != ValidationResult.Success)
-            {
-                ModelState.AddModelError("payment-actual-amount", validPaymentActualAmount.ErrorMessage);
                 _errorService.AddErrors(ModelState.Keys, ModelState);
                 return Page();
             }
@@ -133,27 +109,8 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central
 
             var totalGrant = project.PDGDashboard.RevisedGrant ?? project.PDGDashboard.InitialGrant;
 
-            TotalGrant = (int)totalGrant;
-
             CurrentFreeSchoolName = project.SchoolName;
         }
 
-        private ValidationResult IsValidMoney(object value, string displayName)
-        {
-            if (value is null)
-                return ValidationResult.Success;
-
-            var valueAsDec = (decimal)value;
-
-            if (valueAsDec < 0 || valueAsDec > TotalGrant)
-                return new ValidationResult(string.Format(ValidationConstants.NumberValidationMessage, displayName, 0, TotalGrant));
-
-            if (Math.Round(valueAsDec, 2) != valueAsDec)
-            {
-                return new ValidationResult($"{displayName} must be two decimal places");
-            }
-            
-            return ValidationResult.Success;
-        }
     }
 }