+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Tasks/PDG/Central/AddVariationLetter.cshtml.cs
@@ -2,6 +2,7 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Grants;
+using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks;
 using Dfe.ManageFreeSchoolProjects.Constants;
 using Dfe.ManageFreeSchoolProjects.Logging;
 using Dfe.ManageFreeSchoolProjects.Models;
@@ -16,6 +17,7 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.PDG.Central;
 public class AddVariationLetter(
     IGrantLettersService grantLettersService,
     ILogger<AddVariationLetter> logger,
+    IGetProjectByTaskService getProjectService,
     ErrorService errorService) : PageModel
 {
     public string CurrentFreeSchoolName { get; set; }
@@ -23,13 +25,17 @@ public class AddVariationLetter(
     [BindProperty(SupportsGet = true)] public string ProjectId { get; set; }
 
     [BindProperty(Name = "due-date-variation-letter", BinderType = typeof(DateInputModelBinder))]
-    [Display(Name = "Date")]
+    [Display(Name = "Due date of variation letter")]
+    [Required]
     public DateTime? DueDateOfVariationLetter { get; set; }
 
     [BindProperty] public bool VariationLetterSavedToWorkplaces { get; set; }
 
-    public IActionResult OnGet()
+    public async Task<ActionResult> OnGet()
     {
+        logger.LogMethodEntered();
+        await LoadSchoolName();
+
         return Page();
     }
 
@@ -73,5 +79,11 @@ public class AddVariationLetter(
         return Enum.TryParse<GrantVariationLetter.LetterVariation>(variationNumber, out var variationEnum)
             ? variationEnum
             : GrantVariationLetter.LetterVariation.NotSet;
+    }
+
+    private async Task LoadSchoolName()
+    {
+        var project = await getProjectService.Execute(ProjectId, TaskName.PDG);
+        CurrentFreeSchoolName = project.SchoolName;
     }
 }