+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Startup.cs
@@ -111,6 +111,7 @@ public class Startup
         services.AddScoped<IAddProjectPaymentsService, AddProjectPaymentsService>();
         services.AddScoped<IDeleteProjectPaymentsService, DeleteProjectPaymentsService>();
         services.AddScoped<IGrantLettersService, GrantLettersService>(); 
+        services.AddScoped<IPDGPaymentInfoService, PDGPaymentInfoService>();
 
         services.AddScoped(sp => sp.GetService<IHttpContextAccessor>()?.HttpContext?.Session);
         services.AddSession(options =>
@@ -152,6 +153,7 @@ public class Startup
             MfspOptions mfspOptions = GetTypedConfigurationFor<MfspOptions>();
             client.BaseAddress = new Uri(mfspOptions.ApiEndpoint);
             client.DefaultRequestHeaders.Add("ApiKey", mfspOptions.ApiKey);
+            client.DefaultRequestHeaders.Add("User-Agent", "ManageFreeSchoolProjects/1.0");
         });
 
         services.AddScoped<ErrorService>();
@@ -164,7 +166,6 @@ public class Startup
         });
 
         System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);
-
     }
 
     private void SetupDataprotection(IServiceCollection services)